feat(auth): 사용자 통계 필드(itineraryCount) 증가 로직 추가 및 Firestore 보안 규칙 수정
